feat: adds i18n support (#3)
This update installs the gem 'jekyll-multiple-languages-plugin' which allows for multiple languages in the site.
It also updates the templates in order to support english and Portuguese.
Finally, it translates the first post.

This is a simple update with the goal of validating the language support in production. A few tweaks are still necessary.